+++ api/server.js
@@ -1,11 +1,10 @@
 // モジュールの読み込み
-require('dotenv').config();
-const express = require('express');
-const line = require('@line/bot-sdk');
-const bot = require('../bot');
+import express from 'express';
+import { middleware } from '@line/bot-sdk';
+import 'dotenv/config';
 
-// configの読み込み
-const config = require('../config').index();
+// ファイルの読み込み
+import { index } from '../bot.js';
 
 const PORT = process.env.PORT || 3000;
 const app = express();
@@ -13,7 +12,9 @@ const app = express();
 // /にアクセスがあった時、Deploy succeededと返す
 app.get('/', (req, res) => { res.send('Deploy succeeded'); });
 // /webhookにアクセスがあったとき、bot.jsのindexを呼び出す
-app.post('/webhook', line.middleware(config), bot.index);
+app.post('/webhook', middleware({
+  channelSecret: process.env.channelSecret,
+}), index);
 
 app.listen(PORT);
 console.log(`Server running at ${PORT}`);
