+++ bot.js
@@ -1,22 +1,22 @@
 // モジュール読み込み
-const line = require('@line/bot-sdk');
-const crypto = require('crypto');
-// コンフィグ読み込み
-const config = require('./config').index();
+import line from '@line/bot-sdk';
+import crypto from 'crypto';
 // 各イベントごとの処理をするファイルの読み込み
-const messageFunc = require('./event/message');
-const unsendFunc = require('./event/unsend');
-const postbackFunc = require('./event/postback');
-const joinFunc = require('./event/join');
-const leaveFunc = require('./event/leave');
-const followFunc = require('./event/follow');
-const unfollowFunc = require('./event/unfollow');
-const memberJoinedFunc = require('./event/memberJoined');
-const memberLeftFunc = require('./event/memberLeft');
+import messageFunc from './event/message.js';
+import unsendFunc from './event/unsend.js';
+import postbackFunc from './event/postback.js';
+import joinFunc from './event/join.js';
+import leaveFunc from './event/leave.js';
+import followFunc from './event/follow.js';
+import unfollowFunc from './event/unfollow.js';
+import memberJoinedFunc from './event/memberJoined.js';
+import memberLeftFunc from './event/memberLeft.js';
 
-const client = new line.Client(config);
+const client = new line.Client({
+  channelAccessToken: process.env.channelAccessToken,
+});
 
-exports.index = (req, res) => {
+export const index = (req, res) => {
   console.log(req.body.events);
   // 署名検証
   const signature = crypto.createHmac('sha256', process.env.channelSecret).update(JSON.stringify(req.body)).digest('base64');
@@ -33,70 +33,66 @@ exports.index = (req, res) => {
         case 'message': {
           // メッセージイベントが飛んできた時はmessage.jsのindexを呼び出す
           // 処理結果をmessageに格納
-          message = await messageFunc.index(event, client);
+          message = await messageFunc(event, client);
           break;
         }
         case 'unsend': {
           // 送信取消イベントが飛んできた時はunsend.jsのindexを呼び出す
           // 処理結果をmessageに格納
-          message = unsendFunc.index(event);
+          message = unsendFunc(event);
           break;
         }
         case 'postback': {
           // ポストバックイベントが飛んできた時はpostback.jsのindexを呼び出す
           // 処理結果をmessageに格納
-          message = postbackFunc.index(event);
+          message = postbackFunc(event);
           break;
         }
         case 'join': {
           // 参加イベントが飛んできた時はjoin.jsのindexを呼び出す
           // 処理結果をmessageに格納
-          message = joinFunc.index(event);
+          message = joinFunc();
           break;
         }
         case 'leave': {
           // 退出イベントが飛んできた時はleave.jsのindexを呼び出す
-          leaveFunc.index(event);
+          leaveFunc(event);
           break;
         }
         case 'follow': {
           // フォローイベントが飛んできた時はfollow.jsのindexを呼び出す
-          message = followFunc.index();
+          message = followFunc();
           break;
         }
         case 'unfollow': {
           // フォロー解除イベントが飛んできた時はunfollow.jsのindexを呼び出す
           // 処理結果をmessageに格納
-          unfollowFunc.index(event);
+          unfollowFunc(event);
           break;
         }
         case 'memberJoined': {
           // メンバー参加イベントが飛んできた時はmemberJoined.jsのindexを呼び出す
           // 処理結果をmessageに格納
-          message = memberJoinedFunc.index(event);
+          message = memberJoinedFunc(event);
           break;
         }
         case 'memberLeft': {
           // メンバー退出イベントが飛んできた時はmemberLeft.jsのindexを呼び出す
           // 処理結果をmessageに格納
-          message = memberLeftFunc.index(event);
+          memberLeftFunc(event);
           break;
         }
         default:
           break;
       }
-      // もしmessageがundefinedだった場合
-      if (message === undefined) {
-        // レスポンス
-        res.json('done');
-        // 返信するメッセージが存在する場合
-      } else {
+      // もしmessageがundefinedでなかった場合は返信処理
+      if (message !== undefined) {
         // 返信するメッセージをログに出力
         console.log(`返信メッセージ: ${JSON.stringify(message)}`);
         // メッセージを返信
         client.replyMessage(event.replyToken, message)
-          .then((response) => {
-            res.json(response);
+          .then(() => {
+            console.log('Reply succeeded');
           }).catch((err) => console.log(`${JSON.stringify(message)}\n\n\n${err}`));
       }
     });
@@ -105,4 +101,5 @@ exports.index = (req, res) => {
     // ログに出力
     console.log('署名認証エラー');
   }
+  return res.json('ok');
 };
