--- config.js
@@ -1,4 +1,0 @@
-exports.index = () => ({
-  channelSecret: process.env.channelSecret,
-  channelAccessToken: process.env.channelAccessToken,
-});
