+++ event/unfollow.js
@@ -1,5 +1,5 @@
 // フォロー解除イベントが飛んできた時
-exports.index = (event) => {
+export default (event) => {
   // ログを出力
   console.log(`unfollowされました...\nuserId: ${event.source.userId}`);
 };
