+++ get-content.js
@@ -1,8 +1,8 @@
 // fsライブラリを読み込み
-const fs = require('fs');
+import fs from 'fs';
 
 // ファイルを取得してdownloadディレクトリに出力
-exports.downloadContent = (client, messageId, downloadPath) => client.getMessageContent(messageId)
+export const downloadContent = (client, messageId, downloadPath) => client.getMessageContent(messageId)
   .then((stream) => new Promise((resolve, reject) => {
     const writable = fs.createWriteStream(downloadPath);
     stream.pipe(writable);
